Add beans for keyset and limit-offset strategies 3001
